src/main/java/model/Student.java
package model;


import java.time.LocalDate;

public class Student {

    private String name;

    public LocalDate getDateOfBirth() {
        return dateOfBirth;
    }

    public void setDateOfBirth(LocalDate dateOfBirth) {
        this.dateOfBirth = dateOfBirth;
    }


    private LocalDate dateOfBirth;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

        public static boolean validate(Student s){
            return true;
    }

}

